back_end/compiler/requirements.txt
pyyaml
pyserial
boto3
flask
gunicorn
uuid